Gave a project proper structure and tested all the back end. However,the documentation is in progress
