--- main/src/UseCase/InvalidInputException.java
@@ -1,5 +1,0 @@
-package UseCase;
-
-public class InvalidInputException extends  Exception{
-
-}
