--- main/src/UseCase/OutputBoundary.java
@@ -1,11 +1,0 @@
-package UseCase;
-
-import entity.User;
-import entity.UserList;
-
-public interface OutputBoundary {
-    void showUsers(UserList userList);
-    void outputMessage(String s);
-
-
-}
