+++ src/main/java/CLI/LogInInterface.java
@@ -1,0 +1,15 @@
+package CLI;
+
+import java.io.IOException;
+import java.util.Scanner;
+import interfaceAdapter.Controller;
+import interfaceAdapter.Presenter;
+import useCase.InvalidInputException;
+
+public interface LogInInterface {
+
+    void login(Scanner scanner, Controller controller, Presenter presenter,
+               MainMenuInterface mainMenuInterface,
+               WelcomeMenuInterface welcomeMenuInterface, RegisterInterface registerInterface,
+               ShowUsersInterface showUsersInterface) throws IOException, ClassNotFoundException, InvalidInputException;
+}
