+++ src/main/java/CLI/ShowUsersInterface.java
@@ -1,0 +1,16 @@
+package CLI;
+
+import interfaceAdapter.Controller;
+import interfaceAdapter.Presenter;
+
+import java.io.IOException;
+import java.util.Scanner;
+
+public interface ShowUsersInterface {
+    void showUsers(Scanner scanner, Presenter presenter, Controller controller,
+                   WelcomeMenuInterface welcomeMenuInterface,
+                   RegisterInterface registerInterface,
+                   LogInInterface logInInterface,
+                   MainMenuInterface mainMenuInterface) throws IOException, ClassNotFoundException;
+
+}
