+++ src/main/java/useCase/DataAccess.java
@@ -1,0 +1,12 @@
+package useCase;
+
+import entities.UserList;
+
+import java.io.IOException;
+
+public interface DataAccess {
+
+    UserList importUsers() throws IOException, ClassNotFoundException;
+    void saveUser(UserList userList) throws IOException;
+
+}
