+++ src/main/java/useCase/InvalidInputException.java
@@ -1,0 +1,5 @@
+package useCase;
+
+public class InvalidInputException extends  Exception{
+
+}
