+++ src/main/java/useCase/OutputBoundary.java
@@ -1,0 +1,10 @@
+package useCase;
+
+import entities.UserList;
+
+public interface OutputBoundary {
+    void showUsers(UserList userList);
+    void outputMessage(String s);
+
+
+}
